animation, world_replication, kest, result: 32-bit portability
Found building for wasm32, where size_t is 32 bits and int64_t is long
long: an ambiguous optional literal, a wire component index narrowed
into size_t after its check, a report file that fell back to tmpfile,
a missing <numbers>, and a size assertion that assumed 8-byte
pointers.

+++ modules/world_replication/src/client.cpp
@@ -249,7 +249,8 @@ struct ReplicationClient::State {
             }
             staged.push_back(Staged{.net = kHead->entity.value,
                                     .entity = kMirror->second,
-                                    .component = kHead->component,
+                                    // Checked against the table as it was read.
+                                    .component = static_cast<std::size_t>(kHead->component),
                                     .offset = kOffset});
         }
         if (reader.remaining() != 0) {
